+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 
 # Changelog
 
+## v1.3.2 2018-08-02
+
+- Remove default `target="_blank"` from links
+- Move custom Provider component into app
+
+## v1.3.1 2018-08-02
+
+- Add speaker notes markdown syntax and component
+
+## v1.3.0 2018-08-02
+
+- Add presenter mode with preview of next slide and timer
+
 ## v1.2.3 2018-08-01
 
 - Fix `history.pushState` hash